Fix for onEndReached not getting called (#46)
This PR focuses on fixing onEndReached not getting called. The reason why it sometimes doesn't get called is because of this particular line 

`layoutMeasurement.height + contentOffset.y >= contentSize.height - paddingToBottom`

Where both sides sometimes evaluate to a 6 decimal place number and because of three different measurements (height,offset, total height), comparing them introduces inaccuracy. For example if the 6th decimal point is 0 on left hand side and its 1 on right hand side. Then even if we are at the end, this function will not get called because the fore mentioned line will evaluate to false. To fix this, I have removed the decimal points from the left hand side by using Math.Ceil. This will call the function by removing uncertainties of the decimal points.